Move delete navigation out of hook for invoices
