+++ client/packages/invoices/src/InboundShipment/DetailView/SidePanel/SidePanel.tsx
@@ -8,15 +8,19 @@ import {
   useNotification,
   useDeleteConfirmation,
   useTranslation,
+  RouteBuilder,
+  useNavigate,
 } from '@openmsupply-client/common';
 import { useInbound } from '../../api';
 import { AdditionalInfoSection } from './AdditionalInfoSection';
 import { PricingSection } from './PricingSection';
 import { RelatedDocumentsSection } from './RelatedDocumentsSection';
 import { TransportSection } from './TransportSection';
+import { AppRoute } from '@openmsupply-client/config';
 
 export const SidePanel: FC = () => {
   const t = useTranslation('replenishment');
+  const navigate = useNavigate();
   const { success } = useNotification();
   const { data } = useInbound.document.get();
   const { mutateAsync } = useInbound.document.delete();
@@ -31,7 +35,14 @@ export const SidePanel: FC = () => {
 
   const deleteAction = async () => {
     if (!data) return;
-    await mutateAsync([data]);
+    await mutateAsync([data]).then
+    (() => {
+      navigate(
+        RouteBuilder.create(AppRoute.Replenishment)
+          .addPart(AppRoute.InboundShipment)
+          .build()
+      );
+    });
   };
 
   const onDelete = useDeleteConfirmation({
@@ -72,3 +83,4 @@ export const SidePanel: FC = () => {
     </DetailPanelPortal>
   );
 };
+
