+++ client/packages/invoices/src/InboundShipment/api/hooks/document/useInboundDelete.ts
@@ -1,8 +1,5 @@
-import { AppRoute } from '@openmsupply-client/config';
 import {
-  RouteBuilder,
   useMutation,
-  useNavigate,
   useQueryClient,
 } from '@openmsupply-client/common';
 import { useInboundApi } from '../utils/useInboundApi';
@@ -10,16 +7,10 @@ import { useInboundApi } from '../utils/useInboundApi';
 export const useInboundDelete = () => {
   const queryClient = useQueryClient();
   const api = useInboundApi();
-  const navigate = useNavigate();
 
   return useMutation(api.delete, {
     onSuccess: () => {
       queryClient.invalidateQueries(api.keys.base());
-      navigate(
-        RouteBuilder.create(AppRoute.Replenishment)
-          .addPart(AppRoute.InboundShipment)
-          .build()
-      );
     },
   });
 };
